# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation example",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://test.naufalrafif.com/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.example.com/api",
            "description": "Staging Server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/user/profile": {
            "get": {
                "tags": [
                    "User - Profile"
                ],
                "summary": "Get authenticated user profile",
                "description": "Get authenticated user profile",
                "operationId": "b31b3486e62c3a6170ab1d72fec307cb",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "user_uuid"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "address": {
                                                            "type": "string",
                                                            "example": "Jalan Jalan 1",
                                                            "nullable": true
                                                        },
                                                        "phone_number": {
                                                            "type": "string",
                                                            "example": "081234567890",
                                                            "nullable": true
                                                        },
                                                        "birth_date": {
                                                            "type": "string",
                                                            "example": "1990-01-01",
                                                            "nullable": true
                                                        },
                                                        "gender": {
                                                            "type": "string",
                                                            "example": "male",
                                                            "nullable": true
                                                        },
                                                        "status_account": {
                                                            "type": "string",
                                                            "example": "active",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User - Profile"
                ],
                "summary": "Update authenticated user profile",
                "description": "Update authenticated user profile",
                "operationId": "89a454a77f7fc3fa5fad426ea9cf0ec3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password",
                                        "nullable": true
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jalan Jalan 1",
                                        "nullable": true
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg",
                                        "nullable": true
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890",
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "1990-01-01",
                                        "nullable": true
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male",
                                        "nullable": true
                                    },
                                    "status_account": {
                                        "type": "string",
                                        "example": "active",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile Updated"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "user_uuid"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "address": {
                                                            "type": "string",
                                                            "example": "Jalan Jalan 1",
                                                            "nullable": true
                                                        },
                                                        "phone_number": {
                                                            "type": "string",
                                                            "example": "081234567890",
                                                            "nullable": true
                                                        },
                                                        "birth_date": {
                                                            "type": "string",
                                                            "example": "1990-01-01",
                                                            "nullable": true
                                                        },
                                                        "gender": {
                                                            "type": "string",
                                                            "example": "male",
                                                            "nullable": true
                                                        },
                                                        "status_account": {
                                                            "type": "string",
                                                            "example": "active",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}